+++ philo/inc/structs.h
@@ -6,7 +6,7 @@
 /*   By: joaomigu <joaomigu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/20 00:19:32 by joaomigu          #+#    #+#             */
-/*   Updated: 2025/03/20 10:52:42 by joaomigu         ###   ########.fr       */
+/*   Updated: 2025/03/24 14:47:47 by joaomigu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,8 +27,8 @@ typedef struct s_fork
 typedef struct s_philo
 {
 	int					id;
-	long				meal_counter;
-	long				last_meal_time;
+	unsigned int				meal_counter;
+	unsigned int				last_meal_time;
 	bool				full;
 	t_fork				*first_fork;
 	t_fork				*second_fork;
@@ -39,16 +39,16 @@ typedef struct s_philo
 
 typedef struct s_table
 {
-	long				philo_number;
-	long				time_to_die;
-	long				time_to_eat;
-	long				time_to_sleep;
-	long				time_to_think;
-	long				nbr_limit_meals;
-	long				start_simulation;
+	unsigned int				philo_number;
+	unsigned int				time_to_die;
+	unsigned int				time_to_eat;
+	unsigned int				time_to_sleep;
+	unsigned int				time_to_think;
+	int				nbr_limit_meals;
+	unsigned int				start_simulation;
 	bool				end_simulation;
 	bool				all_threads_ready;
-	long				thread_running_count;
+	unsigned int				thread_running_count;
 	t_mtx				table_mutex;
 	t_mtx				write_mutex;
 	pthread_t			monitor;
